# microservices/api-gateway/openapi.yml
openapi: 3.0.3
info:
  title: POS Multi-Magasins - API Gateway
  description: |
    API Gateway pour le système POS Multi-Magasins avec architecture microservices.
    
    ## Authentification
    
    ### API Keys (Kong Gateway)
    - **Header**: `X-API-KEY`
    - **Valeur**: `pos-microservices-2025`
    
    ### JWT Authentication (Services clients)
    - **Header**: `Authorization`
    - **Format**: `Bearer <jwt_token>`
    - **Endpoint de login**: `POST /api/v1/customers/auth/login`
    
    ## Architecture Microservices
    
    Le système est composé de 7 microservices :
    
    ### Services Métier (extraits du monolithe)
    1. **Product Service** (Port 8001) - Catalogue produits
    2. **Inventory Service** (Port 8002) - Gestion des stocks  
    3. **Sales Service** (Port 8003) - Transactions POS
    4. **Reporting Service** (Port 8004) - Rapports analytiques
    
    ### Services E-commerce (nouveaux)
    5. **Customer Service** (Port 8005) - Comptes clients
    6. **Cart Service** (Port 8006) - Panier d'achats
    7. **Order Service** (Port 8007) - Gestion des commandes
    
    ## Load Balancing
    
    Le Cart Service utilise un load balancer avec 3 instances (algorithme least-connections).
    
    ## Monitoring
    
    - **Prometheus**: Métriques sur `/metrics`
    - **Grafana**: Dashboard Kong Gateway
    - **Health checks**: `/health` sur chaque service
    
  version: "1.0.0"
  contact:
    name: LOG430 Lab 5 Team
    email: lab430@example.com
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: Clé API pour accéder aux microservices via Kong Gateway
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT pour l'authentification des clients
  schemas:
...
    Product:
      type: object
      required: [nom, prix, stock, id_categorie]
      properties:
        id:
          type: integer
          description: ID unique du produit
        nom:
          type: string
          description: Nom du produit
        prix:
          type: number
          format: float
          description: Prix unitaire
        stock:
          type: integer
          description: Quantité en stock
        id_categorie:
          type: integer
          description: ID de la catégorie
        seuil_alerte:
          type: integer
          description: Seuil d'alerte stock
        description:
          type: string
          description: Description du produit
    
    # Modèles Customer Service  
...
    Order:
      type: object
      required: [customer_id, items]
      properties:
        id:
          type: integer
          description: ID de la commande
        customer_id:
          type: integer
          description: ID du client
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
          description: Statut de la commande
        total:
          type: number
          format: float
          description: Total de la commande
        created_at:
          type: string
          format: date-time
          description: Date de création
    OrderItem:
      type: object
      required: [product_id, quantity, unit_price]
      properties:
        product_id:
          type: integer
          description: ID du produit
        quantity:
          type: integer
          minimum: 1
          description: Quantité commandée
        unit_price:
          type: number
          format: float
          description: Prix unitaire
        total_price:
          type: number
          format: float
          description: Prix total de la ligne
    
    # Modèles communs
...
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: Lister les produits
      description: Récupérer la liste des produits avec pagination
      security:
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
          description: Recherche par nom
      responses:
        '200':
          description: Liste des produits
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
    
    post:
      tags: [Products]
      summary: Créer un produit
      description: Ajouter un nouveau produit au catalogue
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Produit créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/v1/products/{product_id}:
    get:
      tags: [Products]
      summary: Récupérer un produit
      description: Obtenir les détails d'un produit spécifique
      security:
        - ApiKeyAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails du produit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produit non trouvé

  # Customer Service Routes
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Créer une commande
      description: Créer une nouvelle commande à partir du panier
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Commande créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/v1/orders/{order_id}:
    get:
      tags: [Orders]
      summary: Récupérer une commande
      description: Obtenir les détails d'une commande
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails de la commande
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  # Inventory Service Routes
  /api/v1/inventory/stores/{store_id}/products:
    get:
      tags: [Inventory]
      summary: Stock par magasin
      description: Récupérer les stocks d'un magasin spécifique
      security:
        - ApiKeyAuth: []
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des stocks du magasin
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    product_id:
                      type: integer
                    product_name:
                      type: string
                    quantity:
                      type: integer
                    alert_threshold:
                      type: integer

  # Sales Service Routes
...